Changed "Confidence" type from int to double
-Changed "Confidence" type from int to double

+++ NetCoreApp3.1/TranslationAPICore/Structs/LanguageDetectResponse.cs
@@ -34,6 +34,6 @@ namespace TranslationAPI.Structs
         /// Detected language confidence
         /// </summary>
         [JsonProperty("confidence")]
-        public int Confidence;
+        public double Confidence;
     }
 }
